backend/requirements.txt
fastapi
uvicorn
pyttsx3
gtts
google-generativeai
google-genai
youtube-transcript-api
PyPDF2
python-multipart
apscheduler
requests
aiofiles
python-dotenv
websockets
